backend/requirements.txt
Django~=3.0.4
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
django-cors-headers==3.2.1
httpie==2.0.0
PyJWT==1.7.1
psycopg2==2.8.4
Pillow==7.1.0